Impl sql delight and it's support for GetProductsApi

+++ src/commonMain/kotlin/onlytrade/app/viewmodel/category/repository/data/db/Category.kt
@@ -1,3 +1,3 @@
 package onlytrade.app.viewmodel.category.repository.data.db
 
-data class Category(val id: Int, val subCategoryId: Int, val name: String)
+data class Category(val id: Long, val subCategoryId: Long, val name: String)
