+++ src/commonMain/kotlin/onlytrade/app/viewmodel/product/add/repository/data/remote/request/AddProductRequest.kt
@@ -7,7 +7,7 @@ import kotlinx.serialization.Transient
 data class AddProductRequest(
     val name: String,
     val description: String,
-    val subcategoryId: Int,
+    val subcategoryId: Long,
     val estPrice: Double,
     @Transient
     val productImages: List<ByteArray>? = null
